requirements.txt
Django
djangorestframework
langchain
langchain_community
langchain_groq
requests
fastapi
uvicorn
python-multipart